requirements.txt
asgiref==3.11.1
attrs==26.1.0
charset-normalizer==3.4.6
dj-database-url==3.1.2
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
gunicorn==25.3.0
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
packaging==26.0
pillow==12.1.1
psycopg2-binary==2.9.11
PyJWT==2.12.1
PyYAML==6.0.3
referencing==0.37.0
reportlab==4.4.10
rpds-py==0.30.0
sqlparse==0.5.5
tzdata==2025.3
uritemplate==4.2.0
